requirements.txt
fastapi==0.126.0
uvicorn[standard]==0.40.0
# uvloop==0.21.0  # Не работает на Windows - используйте asyncio
httptools==0.6.4  # Ускорение HTTP (Linux: полное, Windows: частичное)

SQLAlchemy==2.0.46
asyncpg==0.30.0

python-dotenv==1.1.0
pydantic-settings==2.12.0
python-jose==3.5.0
passlib[bcrypt]==1.7.4

redis==5.2.1
cachetools==5.5.0

telethon==1.41.2
anthropic==0.70.0
openai>=1.0.0

geopy==2.4.1
hdbscan==0.8.37
scikit-learn==1.7.0
requests==2.33.0
httpx==0.28.1

# Безопасность
fastapi-security>=0.3.0
slowapi>=0.1.9
